Refactor

+++ average-waiting-time/Solution.java
@@ -1,6 +1,6 @@
 class Solution {
   public double averageWaitingTime(int[][] customers) {
-    double waitingTimeSum = 0;
+    long waitingTimeSum = 0;
     int chefStartTime = 0;
     for (int[] customer : customers) {
       chefStartTime = Math.max(chefStartTime, customer[0]) + customer[1];
