+++ number-of-1-bits/Solution.java
@@ -1,6 +1,6 @@
 class Solution {
   public int hammingWeight(int n) {
-    long number = (long) n;
+    long number = n;
     if (number < 0) {
       number += 1L << 32;
     }
